fix APF to ignore inner obstacle points, dropping execution time from 2.6s  to 63ms

+++ utils/fast_rrt/src/cuda/graph_best_node.cu
@@ -3,7 +3,7 @@
 #include "../../include/graph.h"
 
 extern __device__ __host__ bool __computeFeasibleForAngle(float3 *frame, int *params, float *classCost, int x, int z, float angle_radians);
-extern __device__ __host__ double getCostCuda(float3 *graphData, long pos);
+extern __device__ __host__ float getCostCuda(float3 *graphData, long pos);
 
 __global__ void __CUDA_KERNEL_findBestNodeWithHeading_bestCost(int3 *graph, float3 *graphData, float3 *frame, int *params, float *classCost, long long searchRadiusSq, int targetX, int targetZ, float targetHeading, long long *bestCost)
 {
