+++ utils/fast_rrt/src/cuda/graph_costs.cu
@@ -3,16 +3,16 @@
 #include "../../include/graph.h"
 
 extern __device__ __host__ int2 getParentCuda(int3 *graph, long pos);
-extern __device__ __host__ double getCostCuda(float3 *graphData, long pos);
+extern __device__ __host__ float getCostCuda(float3 *graphData, long pos);
 extern __device__ __host__ long computePos(int width, int x, int z);
-extern __device__ __host__ double getHeadingCuda(float3 *graphData, long pos);
-extern __device__ __host__ double getFrameCostCuda(float3 *frame, float *classCost, long pos) ;
+extern __device__ __host__ float getHeadingCuda(float3 *graphData, long pos);
+extern __device__ __host__ float getFrameCostCuda(float3 *frame, float *classCost, long pos) ;
 
 
-__device__ __host__ double computeCost(float3 *frame, int3 *graph, float3 *graphData, double *physicalParams, float *classCosts, int width, float goalHeading_rad, long nodePos, double distToParent) {
+__device__ __host__ float computeCost(float3 *frame, int3 *graph, float3 *graphData, double *physicalParams, float *classCosts, int width, float goalHeading_rad, long nodePos, double distToParent) {
     int2 parent = getParentCuda(graph, nodePos);
-    double parentCost = getCostCuda(graphData, computePos(width, parent.x, parent.y));
-    double heading_error_perc = abs(goalHeading_rad - getHeadingCuda(graphData, nodePos)) / physicalParams[PHYSICAL_PARAMS_MAX_STEERING_RAD];
+    float parentCost = getCostCuda(graphData, computePos(width, parent.x, parent.y));
+    float heading_error_perc = abs(goalHeading_rad - getHeadingCuda(graphData, nodePos)) / physicalParams[PHYSICAL_PARAMS_MAX_STEERING_RAD];
     return (getFrameCostCuda(frame, classCosts, nodePos) + distToParent) * (1 + heading_error_perc) + parentCost;
 }
